upgrade to 3.4.1

+++ src/lte/model/lte-spectrum-value-helper.h
@@ -1,3 +1,4 @@
+/* -*- Mode:C++; c-file-style:"gnu"; indent-tabs-mode:nil; -*- */
 /*
  * Copyright (c) 2010 TELEMATICS LAB, DEE - Politecnico di Bari
  *
@@ -22,7 +23,6 @@
 
 #include <ns3/spectrum-value.h>
 
-#include <map>
 #include <vector>
 
 namespace ns3
@@ -90,12 +90,12 @@ class LteSpectrumValueHelper
     /**
      *
      *
-     * \param txBandwidthConf the transmission bandwidth
+     * \param txBandwidthConf the tranmission bandwidth
      * configuration in number of resource blocks
      *
      * \return the nominal channel bandwidth in Hz as per 3GPP TS 36.101
      */
-    static double GetChannelBandwidth(uint16_t txBandwidthConf);
+    static double GetChannelBandwidth(uint8_t txBandwidthConf);
 
     /**
      *
@@ -109,7 +109,7 @@ class LteSpectrumValueHelper
      * configuration. If such SpectrumModel does not exist, it is
      * created.
      */
-    static Ptr<SpectrumModel> GetSpectrumModel(uint32_t earfcn, uint16_t bandwidth);
+    static Ptr<SpectrumModel> GetSpectrumModel(uint32_t earfcn, uint8_t bandwidth);
 
     /**
      * create a spectrum value representing the power spectral
@@ -126,7 +126,7 @@ class LteSpectrumValueHelper
      * Density in W/Hz for each Resource Block
      */
     static Ptr<SpectrumValue> CreateTxPowerSpectralDensity(uint32_t earfcn,
-                                                           uint16_t bandwidth,
+                                                           uint8_t bandwidth,
                                                            double powerTx,
                                                            std::vector<int> activeRbs);
 
@@ -148,7 +148,7 @@ class LteSpectrumValueHelper
      * Density in W/Hz for each Resource Block
      */
     static Ptr<SpectrumValue> CreateTxPowerSpectralDensity(uint32_t earfcn,
-                                                           uint16_t bandwidth,
+                                                           uint8_t bandwidth,
                                                            double powerTx,
                                                            std::map<int, double> powerTxMap,
                                                            std::vector<int> activeRbs);
@@ -169,7 +169,7 @@ class LteSpectrumValueHelper
      * Density in W/Hz for each Resource Block
      */
     static Ptr<SpectrumValue> CreateUlTxPowerSpectralDensity(uint16_t earfcn,
-                                                             uint16_t bandwidth,
+                                                             uint8_t bandwidth,
                                                              double powerTx,
                                                              std::vector<int> activeRbs);
 
@@ -186,7 +186,7 @@ class LteSpectrumValueHelper
      * Density in W/Hz for each Resource Block
      */
     static Ptr<SpectrumValue> CreateNoisePowerSpectralDensity(uint32_t earfcn,
-                                                              uint16_t bandwidth,
+                                                              uint8_t bandwidth,
                                                               double noiseFigure);
 
     /**
